First version of the spec for private documents
In this commit, we describe a spec on how authentication information could be
provided by the vendor to fetch private documents.

Two authentication schemes are described, one with basic auth and another
a flavour of an OAuth2 implementation.

It draws upon the (draft) work of
https://www.w3.org/TR/wot-thing-description/#securityscheme where
possible. It has not fully been taken over yet, as it is clear from
the open issues there is work needed to make WOT properly
RDF-friendly. However, it seems to provide a sound base to extend upon.

Another extension of
http://kanselarij.vo.data.gift/vocabularies/oauth-2.0-session/ has
been introduced too, to support Resource Indicators in OAuth2,
client_id and client_secret
All required to support the initial use case.

http://lblod.data.gift/vocabularies/automatische-melding/
has been extended too with extra predicate
"targetAuthenticationConfiguration" to annotate the link between
authentication data and the submission

A (mini) vocabulary http://lblod.data.gift/vocabularies/security/ has been introduced to link secrets and securityConfiguration