Removed docs-header component
This component was used for styling purposes, giving the introduction a
gray background. This broke the HTML semantics because it separated the
main content section/article from the introduction which should have
been part of the main section/article. It also used old styling which is
not compatible with the new styling. Removing this component also allows
for cleaner docs pages and invites to read from the top instead of
skipping the introduction.