Scroll to top of page on opening a page
Before, the pages would not scroll to the top when using the compact
menu on the bottom of the documentation pages. Sometimes, it would look
like the application wasn't responding because you didn't see the new
page appear when it was scrolled out of view. This is because Ember does
not scroll between subroutes. This addon fixes that.