Review document titles and their routes
Almost all the documents are renamed to be more consistent. Their names
are now in Dutch, the titles in the document are the same as the ones on
the menu and are consolidated to the same structure.
Two new pages were also added where extracted data will end up to
separate the generic Vendor SPARQL API docs from the specific
applications.